+++ src/common/utxobased/info/ecash.ts
@@ -21,8 +21,8 @@ const currencyInfo: EdgeCurrencyInfo = {
 
   // Explorers:
   blockExplorer: 'https://blockchair.com/ecash/block/%s',
-  addressExplorer: 'https://explorer.e.cash/address/ecash:%s',
-  transactionExplorer: 'https://explorer.e.cash/tx/%s',
+  addressExplorer: 'https://blockchair.com/ecash/address/%s',
+  transactionExplorer: 'https://blockchair.com/ecash/tx/%s',
 
   denominations: [
     { name: 'XEC', multiplier: '100', symbol: 'e' },
@@ -65,9 +65,10 @@ const engineInfo: EngineInfo = {
   },
   asBlockbookAddress: asCodec(
     raw => {
-      return asString(raw).split(':')[1]
+      const address = asString(raw)
+      return address.startsWith('ecash:') ? address.split(':')[1] : address
     },
-    address => `ecash:${address}`
+    address => (address.startsWith('ecash:') ? address : `ecash:${address}`)
   )
 }
 
@@ -76,6 +77,7 @@ export const coinInfo: CoinInfo = {
   segwit: false,
   sighash: Psbt.BCH_SIGHASH_ALL,
   coinType: 899,
+  includeCashaddrPrefix: true,
 
   prefixes: {
     messagePrefix: ['\x18Bitcoin Signed Message::\n'],